# src/main/java/com/wallet/interfaces/api/response/ErrorResponse.java
package com.wallet.interfaces.api.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.Instant;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ErrorResponse {
    private Integer statusCode;
    private String message;
    private String path;
    private Instant timestamp;
}