# src/main/resources/openapi/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Wallet Service API
  description: |
    # Digital Wallet Service API
    
    A comprehensive REST API for managing digital wallets, transactions, and users.
    
    ## Features
    
    - **Wallet Management**: Create, retrieve, and manage digital wallets
    - **Transactions**: Deposit, withdraw, and transfer funds
    - **User Management**: User registration and authentication
    - **Transaction History**: Query and filter transaction records
    - **Notifications**: Real-time notifications via Kafka
    
    ## Authentication
    
    All API endpoints (except auth endpoints) require JWT Bearer token authentication.
    Include the token in the Authorization header:
    
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    - 100 requests per second per IP
    - 1000 requests per minute per user
    
    ## Error Codes
    
    | Code | Description |
    |------|-------------|
    | 400  | Bad Request - Invalid input |
    | 401  | Unauthorized - Invalid or missing token |
    | 403  | Forbidden - Insufficient permissions |
    | 404  | Not Found - Resource not found |
    | 409  | Conflict - Resource already exists |
    | 422  | Unprocessable Entity - Validation failed |
    | 500  | Internal Server Error |
    
  version: 1.0.0
  contact:
    name: Wallet Service Team
    email: support@wallet.service
    url: https://wallet.service/support
  license:
    name: Apache 2.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Creates a new user account and returns JWT authentication token
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              name: John Doe
              email: john.doe@example.com
              password: SecurePass123!
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                success: true
                message: User registered successfully
                data:
                  token: eyJhbGciOiJIUzI1NiJ9...
                  tokenType: Bearer
                  expiresIn: 86400000
                statusCode: 200
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: User with this email already exists
                statusCode: 400
        '422':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticates user and returns JWT token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              valid:
                summary: Valid credentials
                value:
                  email: john.doe@example.com
                  password: SecurePass123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: Invalid email or password
                statusCode: 401

  # =============================================================================
  # Wallet Endpoints
  # =============================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint

  # =============================================================================
  # Parameters
  # =============================================================================
...
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          minLength: 8
          maxLength: 100
          format: password
          example: SecurePass123!
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          example: SecurePass123!
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Login successful
        data:
          $ref: '#/components/schemas/JwtToken'
        statusCode:
          type: integer
          example: 200
    JwtToken:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          example: 86400000
          description: Token expiration time in milliseconds
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Wallet not found
        statusCode:
          type: integer
          example: 404
        timestamp:
          type: string
          format: date-time
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        statusCode:
          type: integer
          example: 422
    FieldError:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: Email is required
        rejectedValue:
          type: string
          example: null